+++ test.cpp
@@ -14,7 +14,7 @@ void writeText(string text){
 int main(){
 	cout << "Adjon meg két számot!" << endl;
 
-	int number1,number2;
+	float number1,number2;
 	cout << "Szám 1 : ";
 	cin >> number1;
 	cout << "Szám 2 : ";
